@@ -1,8 +1,7 @@
-﻿namespace EventBus.Events.Interfaces
+﻿namespace EventBus.Events.Interfaces;
+
+public interface IOrderFailedEvent
 {
-    public interface IOrderFailedEvent
-    {
-        public int OrderId { get; set; }
-        public string Reason { get; set; }
-    }
+    public int OrderId { get; set; }
+    public string Reason { get; set; }
 }
