@@ -1,9 +1,8 @@
 ﻿using System;
 using MassTransit;
 
-namespace EventBus.Events.Interfaces
+namespace EventBus.Events.Interfaces;
+
+public interface IPaymentCompletedEvent : CorrelatedBy<Guid>
 {
-    public interface IPaymentCompletedEvent : CorrelatedBy<Guid>
-    {
-    }
 }
