@@ -2,17 +2,16 @@
 using System.Collections.Generic;
 using EventBus.Events.Interfaces;
 
-namespace EventBus.Events
+namespace EventBus.Events;
+
+public class OrderCreatedEvent : IOrderCreatedEvent
 {
-    public class OrderCreatedEvent : IOrderCreatedEvent
+    public OrderCreatedEvent(Guid correlationId)
     {
-        public OrderCreatedEvent(Guid correlationId)
-        {
-            CorrelationId = correlationId;
-        }
-
-        public List<OrderItem> OrderItems { get; set; }
-
-        public Guid CorrelationId { get; }
+        CorrelationId = correlationId;
     }
+
+    public List<OrderItem> OrderItems { get; set; }
+
+    public Guid CorrelationId { get; }
 }
