@@ -1,10 +1,9 @@
 ﻿using EventBus.Events.Interfaces;
 
-namespace EventBus.Events
+namespace EventBus.Events;
+
+public class OrderFailedEvent : IOrderFailedEvent
 {
-    public class OrderFailedEvent : IOrderFailedEvent
-    {
-        public int OrderId { get; set; }
-        public string Reason { get; set; }
-    }
+    public int OrderId { get; set; }
+    public string Reason { get; set; }
 }
