@@ -1,8 +1,7 @@
-namespace EventBus.Events
+namespace EventBus.Events;
+
+public class OrderItem
 {
-    public class OrderItem
-    {
-        public int ProductId { get; set; }
-        public int Count { get; set; }
-    }
+    public int ProductId { get; set; }
+    public int Count { get; set; }
 }
