@@ -1,15 +1,14 @@
 ﻿using System;
 using EventBus.Events.Interfaces;
 
-namespace EventBus.Events
-{
-    public class PaymentCompletedEvent : IPaymentCompletedEvent
-    {
-        public PaymentCompletedEvent(Guid correlationId)
-        {
-            CorrelationId = correlationId;
-        }
+namespace EventBus.Events;
 
-        public Guid CorrelationId { get; }
+public class PaymentCompletedEvent : IPaymentCompletedEvent
+{
+    public PaymentCompletedEvent(Guid correlationId)
+    {
+        CorrelationId = correlationId;
     }
+
+    public Guid CorrelationId { get; }
 }
