@@ -2,18 +2,17 @@
 using System.Collections.Generic;
 using EventBus.Events.Interfaces;
 
-namespace EventBus.Events
+namespace EventBus.Events;
+
+public class PaymentFailedEvent : IPaymentFailedEvent
 {
-    public class PaymentFailedEvent : IPaymentFailedEvent
+    public PaymentFailedEvent(Guid correlationId)
     {
-        public PaymentFailedEvent(Guid correlationId)
-        {
-            CorrelationId = correlationId;
-        }
-
-        public string Reason { get; set; }
-        public List<OrderItem> OrderItems { get; set; }
-
-        public Guid CorrelationId { get; }
+        CorrelationId = correlationId;
     }
+
+    public string Reason { get; set; }
+    public List<OrderItem> OrderItems { get; set; }
+
+    public Guid CorrelationId { get; }
 }
