@@ -2,17 +2,16 @@
 using System.Collections.Generic;
 using EventBus.Events.Interfaces;
 
-namespace EventBus.Events
+namespace EventBus.Events;
+
+public class StockReservedEvent : IStockReservedEvent
 {
-    public class StockReservedEvent : IStockReservedEvent
+    public StockReservedEvent(Guid correlationId)
     {
-        public StockReservedEvent(Guid correlationId)
-        {
-            CorrelationId = correlationId;
-        }
-
-        public List<OrderItem> OrderItems { get; set; }
-
-        public Guid CorrelationId { get; }
+        CorrelationId = correlationId;
     }
+
+    public List<OrderItem> OrderItems { get; set; }
+
+    public Guid CorrelationId { get; }
 }
