@@ -1,9 +1,8 @@
-﻿namespace EventBus.Events
+﻿namespace EventBus.Events;
+
+public class UserCreatedEvent
 {
-    public class UserCreatedEvent
-    {
-        public string UserId { get; set; }
-        public string Email { get; set; }
-        public string UserName { get; set; }
-    }
+    public string UserId { get; set; }
+    public string Email { get; set; }
+    public string UserName { get; set; }
 }
