@@ -3,13 +3,12 @@ using System.Collections.Generic;
 using EventBus.Events;
 using MassTransit;
 
-namespace EventBus.Messages.Interfaces
-{
-    public interface ICompletePaymentMessage : CorrelatedBy<Guid>
-    {
-        public decimal TotalPrice { get; set; }
-        public List<OrderItem> OrderItems { get; set; }
+namespace EventBus.Messages.Interfaces;
 
-        public string CustomerId { get; set; }
-    }
+public interface ICompletePaymentMessage : CorrelatedBy<Guid>
+{
+    public decimal TotalPrice { get; set; }
+    public List<OrderItem> OrderItems { get; set; }
+
+    public string CustomerId { get; set; }
 }
