@@ -1,5 +1,5 @@
 ﻿@page
-@model ErrorModel
+@model WebUI.Pages.ErrorModel
 @{
     ViewData["Title"] = "Error";
 }
