@@ -1,3 +1,2 @@
-﻿@using Discount.WebUI
-@namespace Discount.WebUI.Pages
+﻿@namespace WebUI.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
