@@ -1,3 +1,2 @@
-﻿@using Notification.WebUI
-@namespace Notification.WebUI.Pages
+﻿@namespace WebUI.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
