@@ -1,3 +1,2 @@
-﻿@using Payment.WebUI
-@namespace Payment.WebUI.Pages
+﻿@namespace WebUI.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
