@@ -1,3 +1,2 @@
-﻿@using Subscription.WebUI
-@namespace Subscription.WebUI.Pages
+﻿@namespace WebUI.Pages
 @addTagHelper *, Microsoft.AspNetCore.Mvc.TagHelpers
